+++ stream-access-control/access-control-core/src/main/java/com/backbase/stream/configuration/CustomerAccessGroupConfigurationProperties.java
@@ -1,0 +1,8 @@
+package com.backbase.stream.configuration;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "backbase.stream.customer-access-groups")
+public record CustomerAccessGroupConfigurationProperties(
+    boolean enabled
+) { }