+++ stream-legal-entity/legal-entity-core/src/test/java/com/backbase/stream/ServiceAgreementV2SagaTest.java
@@ -102,6 +102,9 @@ class ServiceAgreementV2SagaTest {
     @Mock
     private PlansService plansService;
 
+    @Mock
+    private CustomerAccessGroupSaga customerAccessGroupSaga;
+
     @Spy
     private final LegalEntitySagaConfigurationProperties legalEntitySagaConfigurationProperties =
         getLegalEntitySagaConfigurationProperties();